+++ satella/coding/concurrent/__init__.py
@@ -3,4 +3,5 @@ from .locked_dataset import LockedDataset
 from .monitor import *
 from .thread import TerminableThread
 
-__all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread']
+__all__ = ['LockedDataset', 'Monitor', 'RMonitor', 'CallableGroup', 'TerminableThread',
+           'MonitorDict', 'MonitorList']